+++ src/components/main/main.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import Main from "./main-component.jsx";
+import Main from "./main.jsx";
 
 Enzyme.configure({
   adapter: new Adapter()
