+++ src/components/main/main.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import Main from "./main-component.jsx";
+import Main from "./main.jsx";
 
 const offers = [
   `Beautiful &amp; luxurious apartment at great location`,
